@@ -1,12 +1,17 @@
 {
   "name": "serverless-questions-api",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "lint": "tslint -p tsconfig.json -c tslint.json",
-    "start": "serverless offline",
-    "deploy": "serverless deploy"
+    "start": "serverless offline start",
+    "deploy": "serverless deploy",
+    "lint": "eslint  --ignore-path .gitignore .",
+    "lint:updated": "pipe-git-updated --ext=js -- eslint --ignore-path .gitignore",
+    "prettier-check": "prettier -c --ignore-path .gitignore \"**/*.{css,html,js,json,md,yaml,yml}\"",
+    "prettier-check:updated": "pipe-git-updated --ext=css --ext=html --ext=js --ext=json --ext=md --ext=yaml --ext=yml -- prettier -c",
+    "prettify": "prettier --write --ignore-path .gitignore \"**/*.{css,html,js,json,md,yaml,yml}\"",
+    "prettify:updated": "pipe-git-updated --ext=css --ext=html --ext=js --ext=json --ext=md --ext=yaml --ext=yml -- prettier --write"
   },
   "keywords": [],
   "author": "",
@@ -15,21 +20,38 @@
     "lint"
   ],
   "devDependencies": {
-    "@types/aws-lambda": "8.10.92",
-    "@types/aws-sdk": "0.0.42",
+    "@aws-sdk/client-dynamodb": "3.214.0",
+    "@aws-sdk/lib-dynamodb": "3.214.0",
+    "@serverless/eslint-config": "5.0.1",
+    "@types/aws-lambda": "8.10.108",
     "@types/lambda-tester": "3.6.1",
-    "@types/node": "^7.0.5",
+    "@types/node": "^18.11.9",
+    "@typescript-eslint/eslint-plugin": "5.43.0",
+    "@typescript-eslint/parser": "5.43.0",
     "cross-env": "7.0.3",
-    "jest": "27.5.1",
+    "eslint": "8.28.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-prettier": "4.2.1",
+    "git-list-updated": "1.2.1",
+    "jest": "29.3.1",
     "lambda-tester": "4.0.1",
-    "serverless-offline": "8.4.0",
-    "serverless-plugin-typescript": "2.1.1",
-    "tslint": "^5.5.0",
-    "tslint-config-standard": "^6.0.1",
-    "typescript": "^2.4.2"
+    "prettier": "2.7.1",
+    "serverless-dynamodb-local": "0.2.40",
+    "serverless-offline": "11.4.0",
+    "serverless-plugin-typescript": "2.1.4",
+    "tslint": "^6.1.3",
+    "tslint-config-standard": "^9.0.0",
+    "typescript": "^4.9.3"
   },
   "dependencies": {
-    "dotenv": "8.6.0",
-    "uuid": "^3.1.0"
+    "dotenv": "16.0.3",
+    "source-map-support": "0.5.21",
+    "uuid": "^9.0.0",
+    "winston": "3.8.2"
+  },
+  "eslintConfig": {
+    "extends": "@serverless/eslint-config/node",
+    "root": true
   }
 }
