@@ -1,32 +1,24 @@
-'use strict';
+import {
+  APIGatewayEvent,
+  APIGatewayProxyHandler,
+  APIGatewayProxyResult,
+  Context,
+} from 'aws-lambda';
+import 'source-map-support/register';
 
-import { DynamoDB } from 'aws-sdk'
+import { listQuestions } from '../repositories/question.repository';
+import { logger } from '../util/logger';
 
-const dynamoDb = new DynamoDB.DocumentClient()
-const params = {
-  TableName: process.env.DYNAMODB_QUESTIONS_TABLE,
-};
+export const list: APIGatewayProxyHandler = async (
+  event: APIGatewayEvent,
+  context: Context
+): Promise<APIGatewayProxyResult> => {
+  logger.defaultMeta = { requestId: context.awsRequestId };
 
-module.exports.list = (event, context, callback) => {
-  // fetch all todos from the database
-  // For production workloads you should design your tables and indexes so that your applications can use Query instead of Scan.
-  dynamoDb.scan(params, (error, result) => {
-    // handle potential errors
-    if (error) {
-      console.error(error);
-      callback(null, {
-        statusCode: error.statusCode || 501,
-        headers: { 'Content-Type': 'text/plain' },
-        body: 'Couldn\'t fetch the todo items.',
-      });
-      return;
-    }
+  const questions = await listQuestions();
 
-    // create a response
-    const response = {
-      statusCode: 200,
-      body: JSON.stringify(result.Items),
-    };
-    callback(null, response);
-  });
+  return {
+    statusCode: 200,
+    body: JSON.stringify(questions),
+  };
 };
